Update Folder Structure and Names + Remove Debug/Unused Stuff
